Clean up integration files; add dot_fast for 3D dot products; change findtransit! to only take Derivatives
