+++ src/integrator/Integrator.jl
@@ -179,15 +179,9 @@ end
 Integrate in the direction of tmax.
 """
 function check_step(t0::T,tmax::T) where T<:AbstractFloat
-    if abs(tmax) > abs(t0)
-        return sign(tmax)
-    else
-        if sign(tmax) != sign(t0)
-            return sign(tmax)
-        else
-            return -1 * sign(tmax)
-        end
-    end
+    if abs(tmax) > abs(t0); return sign(tmax); end
+    if sign(tmax) != sign(t0); return sign(tmax); end
+    return -1 * sign(tmax)
 end
 
 
@@ -198,181 +192,3 @@ for i in ints; include(joinpath(i,"$i.jl")); end
 
 const ints_no_grad = ["ah18","dh17"]
 for i in ints_no_grad; include(joinpath(i,"$(i)_no_grad.jl")); end
-
-"""
-
-Carry out AH18 mapping and calculates both the Jacobian and derivative with respect to the time step.
-"""
-function ah18!(s::State{T},d::Derivatives{T},h::T,pair::Matrix{Bool}) where T<:AbstractFloat
-    zilch = zero(T); uno = one(T); half = convert(T,0.5); two = convert(T,2.0); h2 = half*h; sevn = 7*s.n
-    n = s.n
-    zero_out!(d)
-    fill!(s.dqdt,zilch)
-
-    #drift!(s.x,s.v,s.xerror,s.verror,h2,s.n,s.jac_step,s.jac_error)
-    drift_grad!(s,h2)
-    # Compute time derivative of drift step:
-    @inbounds for i=1:n, k=1:3
-        s.dqdt[(i-1)*7+k] = half*s.v[k,i] + h2*s.dqdt[(i-1)*7+3+k]
-    end
-    #kickfast!(s.x,s.v,s.xerror,s.verror,h/6,s.m,s.n,d.jac_kick,d.dqdt_kick,pair)
-    kickfast!(s,d,h/6,pair)
-    d.dqdt_kick ./= 6 # Since step is h/6
-    # Since I removed identity from kickfast, need to add in dqdt:
-    mul!(d.tmp7n, d.jac_kick, s.dqdt)
-    s.dqdt .+= d.dqdt_kick .+ d.tmp7n #*(d.jac_kick,s.dqdt)
-    # Multiply Jacobian from kick step:
-    if T == BigFloat
-        d.jac_copy .= *(d.jac_kick,s.jac_step)
-    else
-        BLAS.gemm!('N','N',uno,d.jac_kick,s.jac_step,zilch,d.jac_copy)
-    end
-    # Add back in the identity portion of the Jacobian with compensated summation:
-    comp_sum_matrix!(s.jac_step,s.jac_error,d.jac_copy)
-    indi = 0; indj = 0
-    @inbounds for i=1:s.n-1
-        indi = (i-1)*7
-        @inbounds for j=i+1:s.n
-            indj = (j-1)*7
-            if ~pair[i,j]  # Check to see if kicks have not been applied
-                #kepler_driftij_gamma!(s.m,s.x,s.v,s.xerror,s.verror,i,j,h2,d.jac_ij,d.dqdt_ij,true)
-                kepler_driftij_gamma!(s,d,i,j,h2,true)
-                # Pick out indices for bodies i & j:
-                @inbounds for k2=1:sevn, k1=1:7
-                    d.jac_tmp1[k1,k2] = s.jac_step[ indi+k1,k2]
-                    d.jac_err1[k1,k2] = s.jac_error[indi+k1,k2]
-                end
-                @inbounds for k2=1:sevn, k1=1:7
-                    d.jac_tmp1[7+k1,k2] = s.jac_step[ indj+k1,k2]
-                    d.jac_err1[7+k1,k2] = s.jac_error[indj+k1,k2]
-                end
-                # Copy current time derivatives for multiplication purposes:
-                @inbounds for k1=1:7
-                    d.dqdt_tmp1[  k1] = s.dqdt[indi+k1]
-                    d.dqdt_tmp1[7+k1] = s.dqdt[indj+k1]
-                end
-                # Carry out multiplication on the i/j components of matrix:
-                if T == BigFloat
-                    d.jac_tmp2 .= *(d.jac_ij,d.jac_tmp1)
-                else
-                    BLAS.gemm!('N','N',uno,d.jac_ij,d.jac_tmp1,zilch,d.jac_tmp2)
-                end
-                # Add back in the Jacobian with compensated summation:
-                comp_sum_matrix!(d.jac_tmp1,d.jac_err1,d.jac_tmp2)
-                # Copy back to the Jacobian:
-                @inbounds for k2=1:sevn, k1=1:7
-                    s.jac_step[ indi+k1,k2]=d.jac_tmp1[k1,k2]
-                    s.jac_error[indi+k1,k2]=d.jac_err1[k1,k2]
-                end
-                @inbounds for k2=1:sevn, k1=1:7
-                    s.jac_step[ indj+k1,k2]=d.jac_tmp1[7+k1,k2]
-                    s.jac_error[indj+k1,k2]=d.jac_err1[7+k1,k2]
-                end
-                # Add in partial derivatives with respect to time:
-                # Need to multiply by 1/2 since we're taking 1/2 time step:
-                #    BLAS.gemm!('N','N',uno,jac_ij,dqdt_tmp1,half,dqdt_ij)
-                d.dqdt_ij .*= half
-                mul!(d.tmp14, d.jac_ij, d.dqdt_tmp1)
-                d.dqdt_ij .+= d.dqdt_tmp1 .+ d.tmp14#*(d.jac_ij,d.dqdt_tmp1)
-                # Copy back time derivatives:
-                @inbounds for k1=1:7
-                    s.dqdt[indi+k1] = d.dqdt_ij[  k1]
-                    s.dqdt[indj+k1] = d.dqdt_ij[7+k1]
-                end
-            end
-        end
-    end
-    #phic!(s.x,s.v,s.xerror,s.verror,h,s.m,s.n,d.jac_phi,d.dqdt_phi,pair)
-    phic!(s,d,h,pair)
-    #phisalpha!(s.x,s.v,s.xerror,s.verror,h,s.m,two,s.n,d.jac_phi,d.dqdt_phi,pair) # 10%
-    phisalpha!(s,d,h,two,pair)
-    if T == BigFloat
-        d.jac_copy .= *(d.jac_phi,s.jac_step)
-    else
-        BLAS.gemm!('N','N',uno,d.jac_phi,s.jac_step,zilch,d.jac_copy)
-    end
-    # Add in time derivative with respect to prior parameters:
-    #BLAS.gemm!('N','N',uno,jac_phi,dqdt,uno,dqdt_phi)
-    # Copy result to dqdt:
-    mul!(d.tmp7n, d.jac_phi, s.dqdt)
-    s.dqdt .+= d.dqdt_phi .+ d.tmp7n #*(d.jac_phi,s.dqdt)
-    # Add back in the identity portion of the Jacobian with compensated summation:
-    comp_sum_matrix!(s.jac_step,s.jac_error,d.jac_copy)
-    indi=0; indj=0
-    @inbounds for i=s.n-1:-1:1
-        indi=(i-1)*7
-        @inbounds for j=s.n:-1:i+1
-            indj=(j-1)*7
-            if ~pair[i,j]  # Check to see if kicks have not been applied
-                #kepler_driftij_gamma!(s.m,s.x,s.v,s.xerror,s.verror,i,j,h2,d.jac_ij,d.dqdt_ij,false)
-                kepler_driftij_gamma!(s,d,i,j,h2,false)
-                # Pick out indices for bodies i & j:
-                # Carry out multiplication on the i/j components of matrix:
-                @inbounds for k2=1:sevn, k1=1:7
-                    d.jac_tmp1[k1,k2] = s.jac_step[ indi+k1,k2]
-                    d.jac_err1[k1,k2] = s.jac_error[indi+k1,k2]
-                end
-                @inbounds for k2=1:sevn, k1=1:7
-                    d.jac_tmp1[7+k1,k2] = s.jac_step[ indj+k1,k2]
-                    d.jac_err1[7+k1,k2] = s.jac_error[indj+k1,k2]
-                end
-                # Copy current time derivatives for multiplication purposes:
-                @inbounds for k1=1:7
-                    d.dqdt_tmp1[  k1] = s.dqdt[indi+k1]
-                    d.dqdt_tmp1[7+k1] = s.dqdt[indj+k1]
-                end
-                # Carry out multiplication on the i/j components of matrix:
-                if T == BigFloat
-                    d.jac_tmp2 .= *(d.jac_ij,d.jac_tmp1)
-                else
-                    BLAS.gemm!('N','N',uno,d.jac_ij,d.jac_tmp1,zilch,d.jac_tmp2)
-                end
-                # Add back in the Jacobian with compensated summation:
-                comp_sum_matrix!(d.jac_tmp1,d.jac_err1,d.jac_tmp2)
-                # Copy back to the Jacobian:
-                @inbounds for k2=1:sevn, k1=1:7
-                    s.jac_step[ indi+k1,k2]=d.jac_tmp1[k1,k2]
-                    s.jac_error[indi+k1,k2]=d.jac_err1[k1,k2]
-                end
-                @inbounds for k2=1:sevn, k1=1:7
-                    s.jac_step[ indj+k1,k2]=d.jac_tmp1[7+k1,k2]
-                    s.jac_error[indj+k1,k2]=d.jac_err1[7+k1,k2]
-                end
-                # Add in partial derivatives with respect to time:
-                # Need to multiply by 1/2 since we're taking 1/2 time step:
-                #BLAS.gemm!('N','N',uno,jac_ij,dqdt_tmp1,half,dqdt_ij)
-                d.dqdt_ij .*= half
-                mul!(d.tmp14, d.jac_ij, d.dqdt_tmp1)
-                d.dqdt_ij .+= d.dqdt_tmp1 .+ d.tmp14#*(d.jac_ij,d.dqdt_tmp1)
-                # Copy back time derivatives:
-                @inbounds for k1=1:7
-                    s.dqdt[indi+k1] = d.dqdt_ij[  k1]
-                    s.dqdt[indj+k1] = d.dqdt_ij[7+k1]
-                end
-            end
-        end
-    end
-    fill!(d.dqdt_kick,zilch)
-    #kickfast!(x,v,h2,m,n,jac_kick,dqdt_kick,pair)
-    #kickfast!(s.x,s.v,s.xerror,s.verror,h/6,s.m,s.n,d.jac_kick,d.dqdt_kick,pair)
-    kickfast!(s,d,h/6,pair)
-    d.dqdt_kick ./= 6.0 # Since step is h/6
-    # Copy result to dqdt:
-    mul!(d.tmp7n, d.jac_kick, s.dqdt)
-    s.dqdt .+= d.dqdt_kick .+ d.tmp7n#*(d.jac_kick,s.dqdt)
-    # Multiply Jacobian from kick step:
-    if T == BigFloat
-        d.jac_copy .= *(d.jac_kick,s.jac_step)
-    else
-        BLAS.gemm!('N','N',uno,d.jac_kick,s.jac_step,zilch,d.jac_copy)
-    end
-    # Add back in the identity portion of the Jacobian with compensated summation:
-    comp_sum_matrix!(s.jac_step,s.jac_error,d.jac_copy)
-    # Edit this routine to do compensated summation for Jacobian [x]
-    #drift!(s.x,s.v,s.xerror,s.verror,h2,s.n,s.jac_step,s.jac_error)
-    drift_grad!(s,h2)
-    # Compute time derivative of drift step:
-    @inbounds for i=1:n, k=1:3
-        s.dqdt[(i-1)*7+k] += half*s.v[k,i] + h2*s.dqdt[(i-1)*7+3+k]
-    end
-end
