+++ src/transits/timing.jl
@@ -2,7 +2,7 @@
 
 function calc_tt!(s::State{T},intr::Integrator,tt::TransitTiming{T},rstar::T,pair::Matrix{Bool};grad::Bool=true) where T<:AbstractFloat
     n = s.n; ntt_max = tt.ntt;
-    grad ? d = Derivatives(T,s.n) : dT = dTime(T,s.n);
+    d = Derivatives(T,s.n);
     s_prior = deepcopy(s)
     # Define error estimate based on Kahan (1965):
     s2 = zero(T)
@@ -59,10 +59,10 @@ function calc_tt!(s::State{T},intr::Integrator,tt::TransitTiming{T},rstar::T,pai
                 if tt.count[i] <= ntt_max
                     dt0 = -gsave[i]*h/(gi-gsave[i])  # Starting estimate
                     set_state!(s,s_prior) # Set state to step after transit occured
-                    if grad 
+                    if grad
                         dt = findtransit!(tt.ti,i,dt0,s,d,dtdq,intr,pair) # Search for transit time (integrating 'backward')
                     else
-                        dt = findtransit!(tt.ti,i,dt0,s,dT,intr,pair)
+                        dt = findtransit!(tt.ti,i,dt0,s,d,intr,pair)
                     end
                     # Copy transit time and derivatives to TransitTiming structure
                     tt.tt[i,tt.count[i]] = s.t[1] + dt
@@ -95,7 +95,7 @@ function calc_dtdelements!(s::State{T},tt::TransitTiming{T}) where T <: Abstract
     end
 end
 
-function findtransit!(i::Int64,j::Int64,dt0::T,s::State{T},d::Derivatives{T},dtbvdq,intr::Integrator,pair::Array{Bool,2}) where T<:AbstractFloat
+function findtransit!(i::Int64,j::Int64,dt0::T,s::State{T},d::Derivatives{T},dtbvdq::Array{T},intr::Integrator,pair::Array{Bool,2}) where T<:AbstractFloat
     # Computes the transit time, approximating the motion as a fraction of a AH17 step backward in time.
     # Also computes the Jacobian of the transit time with respect to the initial parameters, dtbvdq[1-3,7,n].
     # Initial guess using linear interpolation:
@@ -163,7 +163,7 @@ function findtransit!(i::Int64,j::Int64,dt0::T,s::State{T},d::Derivatives{T},dtb
     end
 end
 
-function findtransit!(i::Int64,j::Int64,dt0::T,s::State{T},dT::dTime{T},intr::Integrator,pair::Array{Bool,2};bv::Bool=false) where T<:AbstractFloat
+function findtransit!(i::Int64,j::Int64,dt0::T,s::State{T},d::Derivatives{T},intr::Integrator,pair::Array{Bool,2};bv::Bool=false) where T<:AbstractFloat
     # Computes the transit time, approximating the motion as a fraction of a AH17 step backward in time.
     # Initial guess using linear interpolation:
 
@@ -186,8 +186,8 @@ function findtransit!(i::Int64,j::Int64,dt0::T,s::State{T},dT::dTime{T},intr::In
         tt1 = dt0
         set_state!(s,s_prior)
         # Advance planet state at start of step to estimated transit time:
-        zero_out!(dT)
-        intr.scheme(s,dT,dt0,pair)
+        zero_out!(d)
+        intr.scheme(s,d,dt0,pair)
         # Compute time offset:
         gsky = g!(i,j,s.x,s.v)
         #  # Compute derivative of g with respect to time:
